Reorganized folder structure for better maintainability
